# target/openapi.json
{"openapi":"3.0.1","info":{"title":"SDS Contact API","description":"Contact Service for SDS Contact application","contact":{"name":"Matt Warman","url":"http://sds.io","email":"matt.warman@sds.io"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"},"version":"V1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/util/email":{"post":{"tags":["util-controller"],"summary":"Send email for the Contact service","operationId":"sendEmail","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Email"}}},"required":true},"responses":{"200":{"description":"Email success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Email"}}}},"400":{"description":"Email problem. Contact SDS admin"}}}},"/contacts/unsubscribe":{"post":{"tags":["contact-controller"],"summary":"Unsubscribes your contact from the selected contact.","operationId":"unsubscribe","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subscriber"}}},"required":true},"responses":{"200":{"description":"Contact unsubscribed","content":{"application/json":{"schema":{"type":"string"}}}},"404":{"description":"subscription not found"},"400":{"description":"Invalid subscriber supplied"}}}},"/contacts/subscriptions":{"post":{"tags":["contact-controller"],"summary":"Retrieves a subscription list for your contact.","operationId":"subscriptions","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subscriber"}}},"required":true},"responses":{"200":{"description":"Contact found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subscription"}}}},"400":{"description":"Invalid subscriber code supplied"},"404":{"description":"Subscription not found"}}}},"/contacts/subscribe":{"post":{"tags":["contact-controller"],"summary":"Subscribes your contact to the selected contact. If not public, the contact will not be available until verified by the subscribed contact.","operationId":"subscribe","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subscriber"}}},"required":true},"responses":{"400":{"description":"Subscription error"},"404":{"description":"subscription not found"},"200":{"description":"Subscription successful","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/contacts/contact":{"post":{"tags":["contact-controller"],"summary":"Creates a new contact","operationId":"newContact","parameters":[{"description":"Contact to add","example":{"contactCode":"JODOE12345","firstName":"John","lastName":"Doe","city":"Cincinnati","state":"OH","details":[{"type":"public","security":"none","detailName":"phone","detailValue":"5133471111"},{"type":"public","security":"none","detailName":"email","detailValue":"john.doe@example.io"}]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Contact"}}},"required":true},"responses":{"400":{"description":"Invalid contact code supplied"},"200":{"description":"Contact created","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/util/states":{"get":{"tags":["util-controller"],"summary":"This method retrieves a list of states for the Contact service","operationId":"getStates","responses":{"400":{"description":"States not found"},"200":{"description":"List of State objects","content":{"application/json":{"schema":{"$ref":"#/components/schemas/State"}}}}}}},"/util/securityTypes":{"get":{"tags":["util-controller"],"summary":"Retrieve a list of security types for the Contact service","operationId":"getSSecurityTypes","responses":{"200":{"description":"List of SecurityType objects","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Security Type not found"}}}},"/util/duplicate/{contactCode}":{"get":{"tags":["util-controller"],"summary":"This method checks contact code for duplicate key","operationId":"checkDuplicateContactCode","parameters":[{"name":"contactCode","in":"path","required":true,"schema":{"type":"string"}},{"description":"Contact contact to check for duplicate","example":{"contactCode":"JODOE12345"}}],"responses":{"200":{"description":"Contact code unique","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Contact code is a duplicate"}}}},"/util/contactTypes":{"get":{"tags":["util-controller"],"summary":"This method retrieves a list of Contact types for the Contact service","operationId":"getContactTypes","responses":{"400":{"description":"States not found"},"200":{"description":"List of Contact Type objects","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/util/categories":{"get":{"tags":["util-controller"],"summary":"Retrieve a list of categories for the Contact service","operationId":"getCategories","responses":{"400":{"description":"States not found"},"200":{"description":"List of Categories","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/contacts/test":{"get":{"tags":["contact-controller"],"operationId":"test","responses":{"200":{"description":"OK","content":{"application/json; charset=UTF-8":{"schema":{"type":"string"}}}}}}},"/contacts/region/{city}/{state}/{type}":{"get":{"tags":["contact-controller"],"summary":"Retrieves a contact by city, state, and contact type.","operationId":"getContactsByRegion","parameters":[{"name":"city","in":"path","required":true,"schema":{"type":"string"}},{"name":"state","in":"path","required":true,"schema":{"type":"string"}},{"name":"type","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Contacts Returned","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Contact"}}}},"404":{"description":"Contacts not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Contact"}}}}}}}},"/contacts/contact/{contactCode}":{"get":{"tags":["contact-controller"],"summary":"Retrieves a contact by contact code.","operationId":"contact","parameters":[{"name":"contactCode","in":"path","description":"Contact code to lookup","required":true,"schema":{"type":"string"},"example":"{'contactCode': 'JODOE123453'}"}],"responses":{"200":{"description":"Contact found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Contact"}}}},"400":{"description":"Invalid contact code supplied"},"404":{"description":"Contact not found"}}}}},"components":{"schemas":{"Email":{"type":"object","properties":{"toEmail":{"type":"string"},"subject":{"type":"string"},"body":{"type":"string"}}},"Subscriber":{"required":["category","contactCode","contactType","subContactCode","subContactType"],"type":"object","properties":{"contactCode":{"type":"string","description":"users contact code","example":"AUSER99999"},"contactType":{"type":"string","description":"subscribers contact type","example":"PUBLIC"},"subContactCode":{"type":"string","description":"subscribers contact code","example":"JODOE12345"},"subContactType":{"type":"string","description":"subscribers contact type","example":"PRIVATE"},"category":{"type":"string","description":"contact category for subscription","example":"VENDOR"},"contactVerified":{"type":"boolean"}}},"Contact":{"required":["city","contactCode","details","firstName","lastName","state"],"type":"object","properties":{"contactCode":{"type":"string","description":"unique contact code","example":"JODOE12345"},"firstName":{"type":"string","description":"first name of the contact","example":"John"},"lastName":{"type":"string","description":"last name of the contact","example":"Doe"},"city":{"type":"string","description":"city of the contact","example":"Cincinnati"},"state":{"type":"string","description":"state abbreviation of the contact","example":"OH"},"details":{"type":"array","description":"A list of contact details","items":{"$ref":"#/components/schemas/Details"}}}},"Details":{"required":["detailName","detailValue","security","type"],"type":"object","properties":{"type":{"type":"string","description":"contact type detail","example":"[\"public\",\"private\",\"work\",\"custom\"]"},"security":{"type":"string","description":"contact type security","example":"none","enum":["none,email,key"]},"detailName":{"type":"string","description":"contact detail name","example":"phone","enum":["title,phone,text, email, address,building, floor,group"]},"detailValue":{"type":"string","description":"contact detail value","example":"5133471111"}},"description":"A list of contact details"},"Subscription":{"required":["category","contactName","subscriberContact"],"type":"object","properties":{"contactName":{"type":"string","description":"contact detail value","example":"5133471111"},"category":{"type":"string","description":"contact detail value","example":"5133471111"},"subscriberContact":{"$ref":"#/components/schemas/Contact"}}},"State":{"required":["abbreviation","name"],"type":"object","properties":{"abbreviation":{"maximum":2,"minimum":2,"type":"string","description":"state abbreviation","example":"OH"},"name":{"type":"string","description":"full state name","example":"Ohio"}}}}}}